moved `ErrorRecovery`
* updated errors
