+++ packages/expo/src/ErrorRecovery/ErrorRecovery.ts
@@ -1,0 +1,12 @@
+import { UnavailabilityError } from 'expo-errors';
+import ExponentErrorRecovery from './ExponentErrorRecovery';
+
+export default {
+  setRecoveryProps(props: { [key: string]: any }): void {
+    if (!ExponentErrorRecovery.setRecoveryProps) {
+      throw new UnavailabilityError('ErrorRecovery', 'setRecoveryProps');
+    }
+
+    return ExponentErrorRecovery.setRecoveryProps(props);
+  },
+};
