+++ packages/expo/src/ErrorRecovery/ExponentErrorRecovery.ts
@@ -1,0 +1,2 @@
+import { NativeModules } from 'react-native';
+export default NativeModules.ExponentErrorRecovery;
